Content review edits

+++ coe-cli/docs/build.ps1
@@ -5,7 +5,7 @@
         .DESCRIPTION
         Generate PDF version of the ebook from docs markdown files
 #>
-param ([string] $branch = "main")
+param ([string] $branch = "main", [boolean] $skipGrammar = $FALSE, [string] $format = "pdf")
 $path = [System.IO.Path]::Combine([System.IO.Path]::GetDirectoryName($MyInvocation.MyCommand.Path))
 Write-Host $path
 if ($branch.Length -gt 0) {
@@ -17,4 +17,14 @@ if ($branch.Length -gt 0) {
     coe ebook generate
 }
 
-docker run -it --rm -v ${path}:/docs cli-mdbook
+if ( $skipGrammar -and $format -eq "docx" ) {
+    docker run -it --rm -e "DOCX=true" -v ${path}:/docs cli-mdbook
+}
+
+if ( $skipGrammar -and $format -eq "pdf" ) {
+    docker run -it --rm -e "SKIP_GRAMMAR=yes" -v ${path}:/docs cli-mdbook
+}
+
+if ( -not $skipGrammar -and $format -eq "pdf" ) {
+    docker run -it --rm -v ${path}:/docs cli-mdbook
+}
